yield_space
